# src/JWT.Generator.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JWT.Generator.API",
    "version": "1.0"
  },
  "paths": {
    "/api/JWTToken/GetToken/{userId}": {
      "get": {
        "tags": [
          "JWTToken"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityToken"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityToken"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityToken"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/JWTToken/GetTokenValue/{userId}": {
      "get": {
        "tags": [
          "JWTToken"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CryptoProviderCache": {
        "type": "object",
        "additionalProperties": false
      },
      "CryptoProviderFactory": {
        "type": "object",
        "properties": {
          "cryptoProviderCache": {
            "$ref": "#/components/schemas/CryptoProviderCache"
          },
          "customCryptoProvider": {
            "$ref": "#/components/schemas/ICryptoProvider"
          },
          "cacheSignatureProviders": {
            "type": "boolean",
            "default": true
          },
          "signatureProviderObjectPoolCacheSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ICryptoProvider": {
        "type": "object",
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "SecurityKey": {
        "type": "object",
        "properties": {
          "keySize": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "keyId": {
            "type": "string",
            "nullable": true
          },
          "cryptoProviderFactory": {
            "$ref": "#/components/schemas/CryptoProviderFactory"
          }
        },
        "additionalProperties": false
      },
      "SecurityToken": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "issuer": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "securityKey": {
            "$ref": "#/components/schemas/SecurityKey"
          },
          "signingKey": {
            "$ref": "#/components/schemas/SecurityKey"
          },
          "validFrom": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "validTo": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}